zlib_sys: fold shared.rs into lib.rs now that posix/win32 are gone
